+++ doorstop/cli/main.py
@@ -54,6 +54,7 @@ def main(args=None):  # pylint: disable=R0915
                           **shared)
     sub.add_argument('prefix',
                      help="document prefix for the new item")
+    sub.add_argument('-l', '--level', help="desired item level (e.g. 1.2.3)")
 
     # Remove subparser
     sub = subs.add_parser('remove',
@@ -92,12 +93,15 @@ def main(args=None):  # pylint: disable=R0915
     sub = subs.add_parser('publish',
                           help="publish a document as text or another format",
                           **shared)
-    sub.add_argument('prefix', help="prefix of document to publish")
-    sub.add_argument('path', nargs='?', help="path of published file")
+    sub.add_argument('prefix', help="prefix of document to publish or 'all'")
+    sub.add_argument('path', nargs='?',
+                     help="path to published file or directory for 'all'")
+    sub.add_argument('-t', '--text', action='store_true',
+                     help="output text (default when no path)")
     sub.add_argument('-m', '--markdown', action='store_true',
-                     help="output Markdown instead of raw text")
+                     help="output Markdown")
     sub.add_argument('-H', '--html', action='store_true',
-                     help="output HTML converted from Markdown")
+                     help="output HTML (default for 'all')")
     sub.add_argument('-w', '--width', type=int,
                      help="limit line width on text output")
 
@@ -187,7 +191,7 @@ def _run(args, cwd, err):  # pylint: disable=W0613
     try:
         tree = build(cwd, root=args.project)
         tree.load()
-        valid = tree.valid()
+        valid = tree.validate()
     except DoorstopError as error:
         logging.error(error)
         return False
@@ -207,13 +211,14 @@ def _run_new(args, cwd, _):
     """
     try:
         tree = build(cwd, root=args.project)
-        document = tree.new(args.root, args.prefix,
-                            parent=args.parent, digits=args.digits)
+        document = tree.new_document(args.root, args.prefix,
+                                     parent=args.parent, digits=args.digits)
     except DoorstopError as error:
         logging.error(error)
         return False
     else:
-        print("created document: {}".format(document.prefix_relpath))
+        print("created document: {} ({})".format(document.prefix,
+                                                 document.relpath))
         return True
 
 
@@ -227,12 +232,12 @@ def _run_add(args, cwd, _):
     """
     try:
         tree = build(cwd, root=args.project)
-        item = tree.add(args.prefix)
+        item = tree.add_item(args.prefix, level=args.level)
     except DoorstopError as error:
         logging.error(error)
         return False
     else:
-        print("added item: {}".format(item.id_relpath))
+        print("added item: {} ({})".format(item.id, item.relpath))
         return True
 
 
@@ -246,12 +251,12 @@ def _run_remove(args, cwd, _):
     """
     try:
         tree = build(cwd, root=args.project)
-        item = tree.remove(args.id)
+        item = tree.remove_item(args.id)
     except DoorstopError as error:
         logging.error(error)
         return False
     else:
-        print("removed item: {}".format(item.id_relpath))
+        print("removed item: {} ({})".format(item.id, item.relpath))
         return True
 
 
@@ -265,13 +270,15 @@ def _run_link(args, cwd, _):
     """
     try:
         tree = build(cwd, root=args.project)
-        child, parent = tree.link(args.child, args.parent)
+        child, parent = tree.link_items(args.child, args.parent)
     except DoorstopError as error:
         logging.error(error)
         return False
     else:
-        print("linked item: {} -> {}".format(child.id_relpath,
-                                             parent.id_relpath))
+        print("linked items: {} ({}) -> {} ({})".format(child.id,
+                                                        child.relpath,
+                                                        parent.id,
+                                                        parent.relpath))
         return True
 
 
@@ -285,13 +292,15 @@ def _run_unlink(args, cwd, _):
     """
     try:
         tree = build(cwd, root=args.project)
-        child, parent = tree.unlink(args.child, args.parent)
+        child, parent = tree.unlink_items(args.child, args.parent)
     except DoorstopError as error:
         logging.error(error)
         return False
     else:
-        print("unlinked item: {} -> {}".format(child.id_relpath,
-                                               parent.id_relpath))
+        print("unlinked items: {} ({}) -> {} ({})".format(child.id,
+                                                          child.relpath,
+                                                          parent.id,
+                                                          parent.relpath))
         return True
 
 
@@ -305,16 +314,16 @@ def _run_edit(args, cwd, _):
     """
     try:
         tree = build(cwd, root=args.project)
-        item = tree.edit(args.id, tool=args.tool, launch=True)
+        item = tree.edit_item(args.id, tool=args.tool, launch=True)
     except DoorstopError as error:
         logging.error(error)
         return False
     else:
-        print("opened item: {}".format(item.id_relpath))
+        print("opened item: {} ({})".format(item.id, item.relpath))
         return True
 
 
-def _run_publish(args, cwd, _):
+def _run_publish(args, cwd, err):
     """Process arguments and run the `doorstop report` subcommand.
 
     @param args: Namespace of CLI arguments
@@ -322,32 +331,55 @@ def _run_publish(args, cwd, _):
     @param err: function to call for CLI errors
 
     """
+    publish_tree = args.prefix == 'all'
     try:
         tree = build(cwd, root=args.project)
-        document = tree.find_document(args.prefix)
+        if publish_tree:
+            documents = [document for document in tree]
+        else:
+            documents = [tree.find_document(args.prefix)]
     except DoorstopError as error:
         logging.error(error)
         return False
     else:
 
+        # Set base arguments
         kwargs = {'ignored': tree.vcs.ignored}
         if args.width:
             kwargs['width'] = args.width
 
+        # Set file extension
         if args.path:
-            ext = os.path.splitext(args.path)[-1]
-        if args.markdown:
+            if publish_tree:
+                ext = '.html'
+            else:
+                ext = os.path.splitext(args.path)[-1]
+        if args.text:
+            ext = '.txt'
+        elif args.markdown:
             ext = '.md'
         elif args.html:
             ext = '.html'
         elif not args.path:
             ext = '.txt'
 
+        # Publish documents
         if args.path:
-            print("publishing {} to {}...".format(document, args.path))
-            report.publish(document, args.path, ext, **kwargs)
+            if publish_tree:
+                print("publishing tree to {}...".format(args.path))
+                for document in documents:
+                    path = os.path.join(args.path, document.prefix + ext)
+                    print("publishing {} to {}...".format(document, path))
+                    report.publish(document, path, ext, **kwargs)
+                if ext == '.html':
+                    report.index(args.path)
+            else:
+                print("publishing {} to {}...".format(documents[0], args.path))
+                report.publish(documents[0], args.path, ext, **kwargs)
         else:
-            for line in report.lines(document, ext, **kwargs):
+            if publish_tree:
+                err("only single documents can be displayed")
+            for line in report.lines(documents[0], ext, **kwargs):
                 print(line)
 
         return True