+++ doorstop/core/document.py
@@ -5,13 +5,14 @@ from itertools import chain
 import logging
 
 from doorstop.core.base import auto_load, auto_save, BaseFileObject
-from doorstop.core.item import Item, split_id
+from doorstop.core.base import BaseValidatable
+from doorstop.core.item import Item, get_id, split_id
 from doorstop import common
-from doorstop.common import DoorstopError, DoorstopWarning, DoorstopInfo
+from doorstop.common import DoorstopError, DoorstopWarning
 from doorstop import settings
 
 
-class Document(BaseFileObject):  # pylint: disable=R0902,R0904
+class Document(BaseValidatable, BaseFileObject):  # pylint: disable=R0902,R0904
 
     """Represents a document directory containing an outline of items."""
 
@@ -53,7 +54,7 @@ class Document(BaseFileObject):  # pylint: disable=R0902,R0904
         if common.VERBOSITY < common.STR_VERBOSITY:
             return self.prefix
         else:
-            return self.prefix_relpath
+            return "{} ({})".format(self.prefix, self.relpath)
 
     def __iter__(self):
         yield from self._iter()
@@ -171,14 +172,15 @@ class Document(BaseFileObject):  # pylint: disable=R0902,R0904
             for filename in filenames:
                 path = os.path.join(dirpath, filename)
                 try:
-                    item = Item(path)
+                    item = Item(path, root=self.root)
                 except DoorstopError:
                     pass  # skip non-item files
                 else:
                     self._items.append(item)
-                    yield item
         # Set meta attributes
         self._itered = True
+        # Yield items
+        yield from self._items
 
     # properties #############################################################
 
@@ -195,6 +197,7 @@ class Document(BaseFileObject):  # pylint: disable=R0902,R0904
 
     @prefix.setter
     @auto_save
+    @auto_load
     def prefix(self, value):
         """Set the document's prefix."""
         self._data['prefix'] = value.strip()
@@ -206,13 +209,6 @@ class Document(BaseFileObject):  # pylint: disable=R0902,R0904
         relpath = os.path.relpath(self.path, self.root)
         return "@{}{}".format(os.sep, relpath)
 
-    # TODO: think of a better name for this property
-    @property
-    @auto_load
-    def prefix_relpath(self):
-        """Get the document's prefix + relative path string."""
-        return "{} ({})".format(self.prefix, self.relpath)
-
     @property
     @auto_load
     def sep(self):
@@ -221,6 +217,7 @@ class Document(BaseFileObject):  # pylint: disable=R0902,R0904
 
     @sep.setter
     @auto_save
+    @auto_load
     def sep(self, value):
         """Set the prefix-number separator to use for new item IDs."""
         # TODO: raise a specific exception for invalid separator characters
@@ -236,6 +233,7 @@ class Document(BaseFileObject):  # pylint: disable=R0902,R0904
 
     @digits.setter
     @auto_save
+    @auto_load
     def digits(self, value):
         """Set the number of digits to use for new item IDs."""
         self._data['digits'] = value
@@ -249,6 +247,7 @@ class Document(BaseFileObject):  # pylint: disable=R0902,R0904
 
     @parent.setter
     @auto_save
+    @auto_load
     def parent(self, value):
         """Set the document's parent document prefix."""
         self._data['parent'] = str(value) if value else ""
@@ -278,8 +277,12 @@ class Document(BaseFileObject):  # pylint: disable=R0902,R0904
 
     # actions ################################################################
 
-    def add(self):
-        """Create a new item for the document and return it."""
+    def add_item(self, level=None):
+        """Create a new item for the document and return it.
+
+        @param level: desired item level
+
+        """
         number = self.next
         logging.debug("next number: {}".format(number))
         try:
@@ -287,7 +290,7 @@ class Document(BaseFileObject):  # pylint: disable=R0902,R0904
         except IndexError:
             level = None
         else:
-            level = last.level[:-1] + (last.level[-1] + 1,)
+            level = level or last.level[:-1] + (last.level[-1] + 1,)
         logging.debug("next level: {}".format(level))
         item = Item.new(self.path, self.root,
                         self.prefix, self.sep, self.digits,
@@ -295,31 +298,33 @@ class Document(BaseFileObject):  # pylint: disable=R0902,R0904
         self._items.append(item)
         return item
 
-    def remove(self, identifier):
+    def remove_item(self, value):
         """Remove an item by its ID.
 
-        @param identifier: item ID
+        @param value: item or its ID
 
         @return: removed Item
 
         @raise DoorstopError: if the item cannot be found
 
         """
+        identifier = get_id(value)
         item = self.find_item(identifier)
         item.delete()
         self._items.remove(item)
         return item
 
-    def find_item(self, identifier, _kind=''):
+    def find_item(self, value, _kind=''):
         """Return an item by its ID.
 
-        @param identifier: item ID
+        @param value: item or its ID
 
         @return: matching Item
 
         @raise DoorstopError: if the item cannot be found
 
         """
+        identifier = get_id(value)
         # Search using the exact ID
         for item in self:
             if item.id.lower() == identifier.lower():
@@ -334,54 +339,75 @@ class Document(BaseFileObject):  # pylint: disable=R0902,R0904
 
         raise DoorstopError("no matching{} ID: {}".format(_kind, identifier))
 
-    def valid(self, tree=None, item_hook=None):
-        """Check the document (and its items) for validity.
-
-        @param tree: Tree containing the document
-        @param item_hook: function to call for custom item validation
-
-        @return: indication that document is valid
-
-        """
-        valid = True
-        # Display all issues
-        for issue in self.issues(tree=tree, item_hook=item_hook):
-            if isinstance(issue, DoorstopInfo):
-                logging.info(issue)
-            elif isinstance(issue, DoorstopWarning):
-                logging.warning(issue)
-            else:
-                assert isinstance(issue, DoorstopError)
-                logging.error(issue)
-                valid = False
-        # Return the result
-        return valid
-
-    def issues(self, tree=None, item_hook=None):
+    def get_issues(self, tree=None, item_hook=None, **_):
         """Yield all the document's issues.
 
-        @param tree: Tree containing the document
+        @param tree: Tree containing the document (tree-level issues)
         @param item_hook: function to call for custom item validation
 
         @return: generator of DoorstopError, DoorstopWarning, DoorstopInfo
 
         """
         logging.info("checking document {}...".format(self))
-        items = list(self)
-        # Check for items
-        if not items:
-            yield DoorstopWarning("no items")
+        # Check levels
+        yield from self._get_issues_level()
         # Check each item
-        for item in items:
+        for item in self:
+            # Check item
             for issue in chain(item_hook(item=item, document=self, tree=tree)
                                if item_hook else [],
-                               item.issues(document=self, tree=tree)):
+                               item.get_issues(document=self, tree=tree)):
                 # Prepend the item's ID to yielded exceptions
                 if isinstance(issue, Exception):
                     yield type(issue)("{}: {}".format(item.id, issue))
+
+    def _get_issues_level(self, items=None):
+        """Yield all the document's issues related to item level."""
+        items = items or self.items
+        # Check for items
+        if not items:
+            return DoorstopWarning("no items")
+        # Check item levels
+        prev = items[0]
+        for item in items[1:]:
+            pid = prev.id
+            plev = prev.level
+            pslev = '.'.join(str(n) for n in plev)
+            nid = item.id
+            nlev = item.level
+            nslev = '.'.join(str(n) for n in nlev)
+            logging.debug("checking level {} to {}...".format(plev, nlev))
+            # Duplicate level
+            if plev == nlev:
+                ids = sorted((pid, nid))
+                msg = "duplicate level: {} ({}, {})".format(pslev, *ids)
+                yield DoorstopWarning(msg)
+            # Skipped level
+            length = min(len(plev), len(nlev))
+            for index in range(length):
+                # Types of skipped levels:
+                #   over: 1.0 --> 1.1
+                #   out: 1.1 --> 3.0
+                #   over and out: 1.1 --> 2.2
+                if (nlev[index] - plev[index] > 1 or  # over or out
+                    (plev[index] != nlev[index] and
+                     index + 1 < length and
+                     nlev[index + 1] not in (0, 1))):  # over and out
+                    msg = "skipped level: {} ({}), {} ({})".format(pslev, pid,
+                                                                   nslev, nid)
+                    yield DoorstopWarning(msg)
+                    break
+            prev = item
 
     def delete(self, path=None):
         """Delete the document and its items."""
         for item in self:
             item.delete()
         super().delete(self.config)
+
+
+# attribute formatters #######################################################
+
+def get_prefix(value):
+    """Get a prefix from a document or string."""
+    return str(value).split(' ')[0]